+++ board/board/src/main/java/com/ys/board/domain/post/service/PostServiceFacade.java
@@ -1,12 +1,12 @@
 package com.ys.board.domain.post.service;
 
-import com.ys.board.domain.post.Post;
-import com.ys.board.domain.post.PostUpdateRequest;
+import com.ys.board.domain.post.model.Post;
+import com.ys.board.domain.post.api.PostUpdateRequest;
 import com.ys.board.domain.post.api.PostCreateRequest;
 import com.ys.board.domain.post.api.PostCreateResponse;
 import com.ys.board.domain.post.api.PostResponse;
 import com.ys.board.domain.post.api.PostResponses;
-import com.ys.board.domain.user.User;
+import com.ys.board.domain.user.model.User;
 import com.ys.board.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
