+++ board/board/src/main/java/com/ys/board/domain/user/api/UserController.java
@@ -1,6 +1,7 @@
 package com.ys.board.domain.user.api;
 
 import com.ys.board.domain.user.service.UserService;
+import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,7 +18,7 @@ public class UserController {
     private final UserService userService;
 
     @PostMapping
-    public ResponseEntity<UserCreateResponse> createUser(@RequestBody UserCreateRequest request) {
+    public ResponseEntity<UserCreateResponse> createUser(@RequestBody @Valid UserCreateRequest request) {
 
         UserCreateResponse createResponse = userService.createUser(request);
