+++ board/board/src/test/java/com/ys/board/domain/post/service/PostServiceSliceTest.java
@@ -7,8 +7,8 @@ import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.verify;
 
 import com.ys.board.common.exception.EntityNotFoundException;
-import com.ys.board.domain.post.Post;
-import com.ys.board.domain.post.PostUpdateRequest;
+import com.ys.board.domain.post.model.Post;
+import com.ys.board.domain.post.api.PostUpdateRequest;
 import com.ys.board.domain.post.api.PostCreateRequest;
 import com.ys.board.domain.post.repository.PostRepository;
 import java.util.Optional;
