+++ board/board/src/test/java/com/ys/board/domain/user/service/UserServiceSliceTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.verify;
 
 import com.ys.board.common.exception.EntityNotFoundException;
 import com.ys.board.domain.user.DuplicateNameException;
-import com.ys.board.domain.user.User;
+import com.ys.board.domain.user.model.User;
 import com.ys.board.domain.user.api.UserCreateRequest;
 import com.ys.board.domain.user.api.UserCreateResponse;
 import com.ys.board.domain.user.repository.UserRepository;
